Work around snap.js state mess
Snap.js is not robust enough to prevent multiple calls to open() or
close(), so we added more checks to prevent these happening even when an
animation is in progress.

Because if we let this through, the bug in question will lead to a
setInterval() call in the library that will hog the CPU forever as the
animation for which its waiting will never finish.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>